refactor: strip down gpu_utils.
