--- src/mrinufft/operators/interfaces/utils/__init__.py
@@ -1,11 +1,0 @@
-"""Utilities for interfaces."""
-
-from .gpu_utils import (
-    get_maxThreadBlock,
-    nvtx_mark,
-)
-
-__all__ = [
-    "get_maxThreadBlock",
-    "nvtx_mark",
-]
